Madlibs generator
must have a file called "madlib.txt" within directory that contains a madlib style sentance (contains $ for nouns, ^ for verbs, and @ for adjectives). The program will subsitute those characters for the user inputted words to create the madlib. Example sentence in input file: "The @ $ loved to ^ every morning before breakfast"